api_rest/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.3
grpcio==1.62.2
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
protobuf==5.26.1
redis==5.0.3
Werkzeug==3.0.2
